third_party/blink/web_tests/external/wpt/webtransport/datagrams.https.any-expected.txt

This is a testharness.js-based test.
[FAIL] Datagrams are echoed successfully
  promise_test: Unhandled rejection with value: object "TypeError: wt.datagrams.createWritable is not a function"
[FAIL] Successfully reading datagrams with BYOB reader.
  promise_test: Unhandled rejection with value: object "TypeError: wt.datagrams.createWritable is not a function"
[FAIL] Reading datagrams with insufficient buffer should be rejected.
  promise_test: Unhandled rejection with value: object "TypeError: wt.datagrams.createWritable is not a function"
[FAIL] Transfer max-size datagram
  promise_test: Unhandled rejection with value: object "TypeError: wt.datagrams.createWritable is not a function"
[FAIL] Fail to transfer max-size+1 datagram, handle PMTUD increases
  promise_test: Unhandled rejection with value: object "TypeError: wt.datagrams.createWritable is not a function"
[FAIL] Sending and receiving datagrams is ready to use before session is established
  promise_test: Unhandled rejection with value: object "TypeError: wt.datagrams.createWritable is not a function"
[FAIL] Datagram's outgoingMaxBufferedDatagrams correctly regulates written datagrams
  promise_test: Unhandled rejection with value: object "TypeError: wt.datagrams.createWritable is not a function"
[FAIL] Datagrams read is less than or equal to the incomingMaxBufferedDatagrams
  promise_test: Unhandled rejection with value: object "TypeError: wt.datagrams.createWritable is not a function"
[FAIL] Datagram MaxAge getters/setters work correctly
  assert_throws_js: function "() => { wt.datagrams.incomingMaxAge = -1; }" did not throw
[FAIL] Datagram MaxBufferedDatagrams getters/setters work correctly
  assert_equals: expected 1 but got 0
Harness: the test ran to completion.
